pyproject.toml
[project]
name = "connecta-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "babel>=2.17.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "pydantic>=2.11.3",
    "pymongo[srv]>=4.12.0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.1",
]
